refactor: make version be an Arc so it could be shared add: traits for block access feat: selections where blocks can be accessed from add: from functions for ComperableValue for easier use add: type abstractions for easier understanding of complex types feat: filter implementation for blocks add: FullBlock for user interaction (and more) add: ChunkPosition for cheaper storing of chunk positions fix: Made Position be whole positions, we already have EntityPosition add: simple implementation of LocalStructure refactor: *started* on separating position functions to utils file

+++ src/loaders/utils.rs
@@ -37,7 +37,7 @@ pub fn get_region_files_in_folder(folder: &PathBuf, dimension_name: &str, region
             continue;
         }
         regions.push(Region {
-            position: Position::new(dimension_name, region_x.unwrap() as f32, 0f32, region_z.unwrap() as f32),
+            position: Position::new(dimension_name, region_x.unwrap(), 0, region_z.unwrap()),
             path: file.path().to_path_buf(),
             region_type: region_type.clone() // bad maybe?
         })
